+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Output `.dylib` on MacOS - #66
 - Update to `kserd` 0.3
 - **Breaking Change:** Renaming of `to_*` to `into_*` in various places.
+- **Breaking Change:** Removed `SourceCode::new` in-lieu of `Default` impl.
 
 ## 0.13.0
 - Restructure of repository